feat: implement POST /rentals endpoint and fix type mismatches

+++ src/main/java/com/chatop/chatop/dtos/RentalDto.java
@@ -3,19 +3,19 @@ package com.chatop.chatop.dtos;
 import java.util.Date;
 
 public class RentalDto {
-    private Long id;
+    private Integer id;
     private String name;
     private Integer surface;
     private Integer price;
     private String picture;
     private String description;
-    private Long owner_id;
+    private Integer owner_id;
     private Date created_at;
     private Date updated_at;
 
     // Getters et Setters
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; }
+    public Integer getId() { return id; }
+    public void setId(Integer id) { this.id = id; }
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
     public Integer getSurface() { return surface; }
@@ -26,8 +26,8 @@ public class RentalDto {
     public void setPicture(String picture) { this.picture = picture; }
     public String getDescription() { return description; }
     public void setDescription(String description) { this.description = description; }
-    public Long getOwner_id() { return owner_id; }
-    public void setOwner_id(Long owner_id) { this.owner_id = owner_id; }
+    public Integer getOwner_id() { return owner_id; }
+    public void setOwner_id(Integer owner_id) { this.owner_id = owner_id; }
     public Date getCreated_at() { return created_at; }
     public void setCreated_at(Date created_at) { this.created_at = created_at; }
     public Date getUpdated_at() { return updated_at; }
